requirements.txt
aiohttp==3.6.2
alembic==1.4.3
amqp==2.6.1
apache-airflow==1.10.12
apispec==3.3.2
argcomplete==1.12.1
async-timeout==3.0.1
attrs==19.3.0
Authlib==0.14.3
Babel==2.8.0
billiard==3.6.3.0
cached-property==1.5.2
cachelib==0.1.1
cachetools==4.1.1
cattrs==1.0.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.35
cryptography==3.1.1
dataclasses==0.7
dateparser==0.7.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
dnspython==2.0.0
docker==4.2.2
docutils==0.16
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==3.0.1
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OAuthlib==0.9.6
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
flower==0.9.5
fsspec==0.8.0
funcsigs==1.0.2
future==0.18.2
gcsfs==0.7.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.1.0
google-cloud-bigquery-storage==2.0.0
google-cloud-bigtable==1.5.1
google-cloud-container==1.0.1
google-cloud-core==1.4.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.0
google-cloud-speech==1.3.2
google-cloud-storage==1.31.2
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphviz==0.14.2
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
grpcio-gcp==0.2.2
gunicorn==20.0.4
httplib2==0.18.1
humanize==3.0.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
json-merge-patch==0.2
jsonschema==3.2.0
kombu==4.6.11
kubernetes==11.0.0
lazy-object-proxy==1.5.1
libcst==0.3.12
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
multidict==4.7.6
mypy-extensions==0.4.3
natsort==7.0.1
numpy==1.19.2
oauthlib==2.1.0
pandas==0.25.3
pandas-gbq==0.14.0
pendulum==1.4.4
pip==19.0.2
prison==0.1.3
prometheus-client==0.8.0
proto-plus==1.10.1
protobuf==3.13.0
psutil==5.7.2
psycopg2-binary==2.8.5
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybuilder==0.12.7
pycparser==2.20
pydata-google-auth==1.1.0
Pygments==2.7.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.10.11
requests==2.24.0
requests-oauthlib==1.1.0
rsa==4.6
setproctitle==1.1.10
setuptools==49.6.0
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
tornado==5.1.1
typing==3.7.4.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
websocket-client==0.57.0
Werkzeug==0.16.1
wheel==0.35.1
WTForms==2.3.3
yarl==1.6.0
zipp==3.3.0
zope.deprecation==4.4.0